+++ Editor/ContextProperties.cs
@@ -40,6 +40,4 @@ namespace SOSXR.EditorSpice
             }
         }
     }
-    
-    
 }