+++ Editor/PivotUtilities.cs
@@ -1,0 +1,139 @@
+using UnityEditor;
+using UnityEngine;
+
+
+// From: [talecrafter](https://gist.github.com/talecrafter/519e260d93dbf236484acfe625faa1dc)
+namespace SOSXR.EditorSpice
+{
+    public static class PivotUtilities
+    {
+        [MenuItem("SOSXR/Pivot/Create Pivot", false, 0)]
+        private static void CreatePivotObject()
+        {
+            if (Selection.activeGameObject != null)
+            {
+                var pivot = CreatePivotObject(Selection.activeGameObject);
+                Selection.activeGameObject = pivot;
+            }
+        }
+
+
+        [MenuItem("SOSXR/Pivot/Create Pivot (Local Zero)", false, 0)]
+        private static void CreatePivotObjectAtParentPos()
+        {
+            if (Selection.activeGameObject != null)
+            {
+                var pivot = CreatePivotObjectAtParentPos(Selection.activeGameObject);
+                Selection.activeGameObject = pivot;
+            }
+        }
+
+
+        [MenuItem("SOSXR/Pivot/Delete Pivot", false, 0)]
+        private static void DeletePivotObject()
+        {
+            GameObject objSelectionAfter = null;
+
+            if (Selection.activeGameObject != null)
+            {
+                if (Selection.activeGameObject.transform.childCount > 0)
+                {
+                    objSelectionAfter = Selection.activeGameObject.transform.GetChild(0).gameObject;
+                }
+                else if (Selection.activeGameObject.transform.parent != null)
+                {
+                    objSelectionAfter = Selection.activeGameObject.transform.parent.gameObject;
+                }
+
+                DeletePivotObject(Selection.activeGameObject);
+
+                Selection.activeGameObject = objSelectionAfter;
+            }
+        }
+
+
+        private static GameObject CreatePivotObjectAtParentPos(GameObject current)
+        {
+            if (current == null)
+            {
+                return null;
+            }
+
+            var siblingIndex = current.transform.GetSiblingIndex();
+
+            var newObject = new GameObject("Pivot");
+            newObject.transform.SetParent(current.transform.parent);
+
+            newObject.transform.localPosition = Vector3.zero;
+            newObject.transform.localScale = Vector3.one;
+            newObject.transform.localRotation = Quaternion.identity;
+
+            newObject.transform.SetSiblingIndex(siblingIndex);
+
+            current.transform.SetParent(newObject.transform);
+
+            return newObject;
+        }
+
+
+        private static GameObject CreatePivotObject(GameObject current)
+        {
+            if (current == null)
+            {
+                return null;
+            }
+
+            var siblingIndex = current.transform.GetSiblingIndex();
+
+            var newObject = new GameObject("Pivot");
+            newObject.transform.SetParent(current.transform.parent);
+
+            newObject.transform.position = current.transform.position;
+            newObject.transform.localScale = current.transform.localScale;
+            newObject.transform.rotation = current.transform.rotation;
+
+            newObject.transform.SetSiblingIndex(siblingIndex);
+
+            current.transform.SetParent(newObject.transform);
+
+            return newObject;
+        }
+
+
+        private static GameObject DeletePivotObject(GameObject current)
+        {
+            var parent = current.transform.parent;
+            var childrenCount = current.transform.childCount;
+            var siblingIndex = current.transform.GetSiblingIndex();
+
+            var children = new Transform[childrenCount];
+
+            for (var i = 0; i < childrenCount; i++)
+            {
+                children[i] = current.transform.GetChild(i);
+            }
+
+            for (var i = 0; i < childrenCount; i++)
+            {
+                children[i].SetParent(parent);
+                children[i].SetSiblingIndex(siblingIndex + i);
+            }
+
+            if (Application.isPlaying)
+            {
+                GameObject.Destroy(current);
+            }
+            else
+            {
+                GameObject.DestroyImmediate(current);
+            }
+
+            if (children.Length > 0)
+            {
+                return children[0].gameObject;
+            }
+
+            return null;
+        }
+    }
+}