+++ Samples~/Samples/FadeBoolDemo.cs
@@ -5,6 +5,5 @@ namespace SOSXR.EditorSpice.Samples
 {
     public class FadeBoolDemo : MonoBehaviour
     {
-        
     }
 }